Factor out common code between inspect and run-local commands.
